Move createLockTable into LockTableService
