+++ atlasdb-cassandra-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/SchemaMutationLockTest.java
@@ -73,7 +73,7 @@ public class SchemaMutationLockTest {
         CassandraKeyValueServiceConfigManager simpleManager = CassandraKeyValueServiceConfigManager.createSimpleManager(quickTimeoutConfig);
         ConsistencyLevel writeConsistency = ConsistencyLevel.EACH_QUORUM;
         CassandraClientPool clientPool = new CassandraClientPool(simpleManager.getConfig());
-        LockTableService lockTableService = new LockTableService();
+        LockTableService lockTableService = new LockTableService(simpleManager, clientPool);
 
         schemaMutationLock = new SchemaMutationLock(supportsCas, simpleManager, clientPool, writeConsistency, lockTableService);
     }
